requirements.txt
requests
beautifulsoup4
selenium
webdriver-manager
undetected-chromedriver
Flask
Flask-Login
Flask-SQLAlchemy
pythainlp
psycopg2-binary 
gunicorn
python-dotenv
openai
Flask
pandas
openpyxl